rework the pll clock to mirror the actual hardware better
The included muxing in the previous version felt artificial and also
did not take the second xin32k source into account at all.
Now we create two stacked clocks one for the actual pll and on top
of it the mux-clock for handling the source selection for the actual
pll output.