clk: rockchip: add initcall to set clk defaults after syscons are available
PLLs on Rockchip platforms report their locking state in an external register
situated in the "General Register Files" which is provided through a syscon
device.

When the initial clk init runs, this syscon is of course not yet available,
making it impossible to set PLLs to other frequencies through the assigned-rate
property of the clock-controller node.

Syscon devices are initialized through a postcore initcall, add an arch_initcall
to rerun the rockchip specific clock initalization when the GRF is available.

As the clock init already runs two times (through of_clk_add_provider and
of_clk_init), a third time shouldn't hurt to much and in the best case
wouldn't change any settings at all.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>