FROMLIST: libata: add ledtrig support
This adds a LED trigger for each ATA port indicating disk activity.

As this is needed only on specific platforms (NAS SoCs and such),
these platforms should define ARCH_WANTS_LIBATA_LEDS if there
are boards with LED(s) intended to indicate ATA disk activity and
need the OS to take care of that.
In that way, if not selected, LED trigger support not will be
included in libata-core and both, codepaths and structures remain
untouched.

I'm currently deploying this for the oxnas target in OpenWrt
https://dev.openwrt.org/changeset/43675/

v2: rebased to kernel/git/tj/libata.git
    plus small corrections and comments added

[Aditya Prayoga:
 * Port forward
 * Change ATA_LEDS default to no
 * Reduce performance impact by moving ata_led_act() call from
    ata_qc_new() to ata_qc_complete()]

v3: Heiko Stuebner:
- remove ARCH_WANT_LIBATA_LEDS (a lot of arches can use SATA ports)
- led-trigger-event takes numerical values
- remove inline function (Andrew Lunn)
- move from #ifdef to IS_ENABLED (Andrew Lunn)

Signed-off-by: Daniel Golle <daniel@makrotopia.org>
URL: https://patchwork.ozlabs.org/patch/420733/
Signed-off-by: Aditya Prayoga <aditya@kobol.io>
Link: https://lore.kernel.org/r/1537328730-9156-2-git-send-email-aditya@kobol.io
Signed-off-by: Heiko Stuebner <heiko@sntech.de>