arm64: dts: rockchip: use correct pinctrl for syr82x on rk3399 firefly
On the firefly the vsel pins of the syr82x regulators are connected to
soc gpios, making them software controllable.
Right now the pins do not get configured at all, so cpu_b_sleep_h is
configured as input with a pull up while the gpu_sleep_h has a pull-down.

In contrast the pin names indicate both should be high in their sleep
state and low in normal operation and the fcs,suspend-voltage-selector
does mirror this setting, so enable the already existing pinctrl
entries for them and also fix their naming to match the schematics.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>