ARM: dts: rockchip: use non-secure timer as broadcast on rk3288
The rk3288 soc has 2 separate timer blocks one with 6 channels and
another one with 2 channels that can be set as secure.

Firmware variants using psci might want to make use of that feature
making both channels of the secure timer inaccessible for a kernel
running in non-secure mode, which affects the timer currently used
as broadcast.

Therefore move over to using the other (still unused) timer block
for that. The timer clock in the original node was also wrong,
as it was pointing to the 6-channel block already, while the clock
of the 2ch block gets controlled from the secure-grf.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>