drm/rockchip: dw-dp: call phy init and exit functions
The dw-dp connected to for example the rk3588's typec phys cannot really
detect the monitor presence on its own. Instead the phy manually sets the
hotplug state according to the detected usb altmode.

For this the phy needs to write to the dp-section of the vo0-grf.
When dp clocks are not running this will cause an SERROR exception, so
the phy should only do that when the dp controller has probed.

So use phy_init and phy_exit calls from the phy-API to tell the phy
something is using it.

Signed-off-by: Heiko Stuebner <heiko.stuebner@cherry.de>