TSD: arm64: set SERIAL_8250_RUNTIME_UARTS to 10 on jaguar-rk3588_defconfig
CONFIG_SERIAL_8250_RUNTIME_UARTS defaults to 4 ... the number of ports
initialized by default at boot. SERIAL_8250_NR_UARTS on the other hand
specifies the max number of ports in the system and is set to 10 .

In theory when the number of needed uarts then exceeds NR_UARTS, the 8250
driver should then extend its array of ports with the new one.

The problem here lies in the port-selection mechanism.

The 8250 at first tries to match the port against the actual entry in the
list, which includes respecting the alias.
If that fails, the next mechanism is to find an unused spot inside the
list of created ports and finally if that doesn't work the final solution
is to create a new entry at the end of the array and increase the nr_uarts
number.

Of this 3 ways only the very first one respects the alias, if the uart
creating reaches the others, the number will be somewhat random, leading
to for example:
[    0.151681] fd890000.serial: ttyS0 at MMIO 0xfd890000 (irq = 39, base_baud = 1500000) is a 16550A
[    0.152049] feb50000.serial: ttyS2 at MMIO 0xfeb50000 (irq = 40, base_baud = 1500000) is a 16550A
[    1.849007] feb60000.serial: ttyS3 at MMIO 0xfeb60000 (irq = 41, base_baud = 1500000) is a 16550A
[    1.859412] feba0000.serial: ttyS1 at MMIO 0xfeba0000 (irq = 42, base_baud = 1500000) is a 16550A

instead of the expected
[    0.190115] fd890000.serial: ttyS0 at MMIO 0xfd890000 (irq = 39, base_baud = 1500000) is a 16550A
[    0.191532] feb50000.serial: ttyS2 at MMIO 0xfeb50000 (irq = 40, base_baud = 1500000) is a 16550A
[    0.194653] feb60000.serial: ttyS3 at MMIO 0xfeb60000 (irq = 41, base_baud = 1500000) is a 16550A
[    0.195936] feba0000.serial: ttyS7 at MMIO 0xfeba0000 (irq = 42, base_baud = 1500000) is a 16550A

because ttyS7 is placed in the hole of the unused ttyS1.

Past discussions already revealed a lot of different opinions when it
comes to aliases and their use and possible holes in alias numbering.

So instead, just set the RUNTIME_UARTS to 10 for us.

Relates-to: JAG-215
Signed-off-by: Heiko Stuebner <heiko.stuebner@cherry.de>