arm64: dts: rockchip: use clock-generator for pcie refclk on rock3a
The rock3a uses a clock-generator chip to create the reference clocks
needed both in the pcie controller and the pcie device. It currently
declares a phy supply in the pcie30-phy to make sure that reference clock
is turned on.

This violates the binding for the phy and this reference clock should
instead be declared in the pcie controller. So use the clock-generator
binding to describe the reference clock and fix the underlying
dtbscheck warning.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>