BACKPORT: drm/panthor: add DRM fdinfo support
Drawing from the FW-calculated values in a previous commit, we can increase
the numbers for an open file by collecting them from finished jobs when
updating their group synchronisation objects.

Display of fdinfo key-value pairs is governed by a bitmask that is by
default unset in the present commit, and supporting manual toggle of it
will be the matter of a later commit.

Signed-off-by: Adrián Larumbe <adrian.larumbe@collabora.com>
Reviewed-by: Steven Price <steven.price@arm.com>
Reviewed-by: Boris Brezillon <boris.brezillon@collabora.com>
Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240923230912.2207320-4-adrian.larumbe@collabora.com
(cherry picked from commit e16635d88fa07ba5801aa9e57ad7fe3c053234e4)

[Backport notes: Both the fop_flags and show_fdinfo elements where
introduced around the same time, fop_flags as fix for 6.12 and other as
new feature. The fix is done the same as Dave did for the main drm tree:
https://gitlab.freedesktop.org/drm/kernel/-/commit/aa628ebb069e37ec5297361a952f7b7b9d980585 ]
Signed-off-by: Heiko Stuebner <heiko.stuebner@cherry.de>