FROMLIST: dt-bindings: leds: Document netdev trigger netdev-trigger-mode property
Document netdev trigger specific netdev-trigger-mode property which
is used to configure the netdev trigger mode flags. Those mode flags
define events on which the LED acts upon when the hardware offload is
enabled. This is traditionally configured via sysfs, but that depends
on udev rules which are available either too late or never in case of
non-Linux systems.

For each LED with linux,default-trigger = "netdev" described in DT, this
optional netdev-trigger-mode property supplies the default configuration
of the PHY LED mode via DT. This property should be set to a subset of
TRIGGER_NETDEV_* flags.

Signed-off-by: Marek Vasut <marex@denx.de>
Link: https://lore.kernel.org/r/20250113002346.297481-1-marex@denx.de
Signed-off-by: Heiko Stuebner <heiko@sntech.de>