arm64: dts: rockchip: add overlay for qnap-ts233 device revision
It seems there exists a revised board/backplane version called
q0aj0_q0am0_12_11, that adds gpios for detecting the presence of
individual hard-drives and for controlling power to the individual
slots.

The presence detection is not really necessary for functionality and
there are also no existing bindings for this.
So add them as gpio hogs to at least document them.

FIXME: need always-on regulators for the slot power

Signed-off-by: Heiko Stuebner <heiko@sntech.de>