arm64: dts: rockchip: connect mac-address nvmem cell to gmac on qnap-ts233/ts433
Connect the relevant eeprom cells to the gmac to retrieve the mac address
from.

While both boards use gmac0 as their 1G ethernet, the mac cell used differs.
Probably because the ts433 also brings a 2.5G PCIe ethernet controller which
likely is designated as "first" ethernet, but has a separate on-controller
memory for its mac address.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>