UPSTREAM: mfd: qnap-mcu: Add driver data for TS233 variant
Add the TS233 compatible and affiliated driver data to qnap-mcu.
The TS233 is mostly similar to the TS433, except not having any of the
PCIe components, so there are only 2 drives.

The fan pwm-limits from the vendor-configuration also are the same
as for the ts433 variant.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Link: https://lore.kernel.org/r/20250730173423.1878599-3-heiko@sntech.de
Signed-off-by: Lee Jones <lee@kernel.org>
(cherry picked from commit 25ccf3e4883234de2623e5cae973c0e90e71524e)
Signed-off-by: Heiko Stuebner <heiko@sntech.de>