arm64: dts: rockchip: describe the NPU power-domain on RK356x
The NPU power-domain has already been described in the dt-binding, but
probably due to not being usable by anything has not been described
in the actual RK356x devicetree.

So do that now and also include a label, because the NPU of the RK356x has
an external supply. This supply also affects the power domain of the NPU,
not just the NPU device nodes themselves.

Since correctly modelled boards will want the power domain to be aware
of the regulator so that it doesn't always have to be on, add a label to
the NPU power domain node so board files can reference it.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>