UPSTREAM: arm64: dts: rockchip: move common qnap tsx33 parts to dtsi
The NAS series based around the rk3568 contains a number of models with
1-4 drives, that reuse most of the board structure.

Therefore move the shared parts to a dtsi, to be included by the devices.

As the smallest device is the 1-bay TS133, keep everything > slot1 in
the individual devicetree.

Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Link: https://patch.msgid.link/20251112214206.423244-4-heiko@sntech.de
(cherry picked from commit 35202606ea08b9ec2b1386ac937a729e25d5beb0)
Signed-off-by: Heiko Stuebner <heiko@sntech.de>